Add error state when the client encounters an unrecoverable problem

+++ lib/drivers/stdio.ts
@@ -15,7 +15,15 @@ class StdioDriver implements IDriver {
     private _outstandingRequests = new Map<number, AsyncSubject<any>>();
     private _projectPath: string;
     private _serverPath: string;
-    public currentState: DriverState = DriverState.Disconnected;
+    private _currentState: DriverState = DriverState.Disconnected;
+    public get currentState() { return this._currentState; }
+    public set currentState(value) {
+        // Prevent the client from leaving the error state
+        // Error means bad things, like mono is working right
+        if (this._currentState !== DriverState.Error) {
+            this._currentState = value;
+        }
+    }
     public id: string;
 
     constructor({projectPath, debug, serverPath}: IDriverOptions) {
@@ -25,7 +33,7 @@ class StdioDriver implements IDriver {
         this._connectionStream.subscribe(state => this.currentState = state);
     }
 
-    public get serverPath() { return this._serverPath ; }
+    public get serverPath() { return this._serverPath; }
     public get projectPath() { return this._projectPath; }
 
     private _commandStream = new Subject<OmniSharp.Stdio.Protocol.ResponsePacket>();
@@ -55,13 +63,17 @@ class StdioDriver implements IDriver {
         });
         rl.on('line', (data) => this.handleData(data));
 
-        this._process.on('close', () => this.disconnect());
+        if (this._process.pid)
+            this.id = this._process.pid.toString();
+
         this._process.on('error', (data) => this.serverErr(data));
-        this.id = this._process.pid.toString();
+        this._process.on('close', () => this.disconnect());
     }
 
     private serverErr(data) {
         var friendlyMessage = this.parseError(data);
+        this.currentState = DriverState.Error;
+        this._process = null;
 
         this._eventStream.onNext({
             Type: "error",
@@ -83,7 +95,7 @@ class StdioDriver implements IDriver {
 
     public disconnect() {
         this._connectionStream.onNext(DriverState.Disconnected);
-        if (this._process != null) {
+        if (this._process != null && this._process.pid) {
             this._process.kill("SIGTERM");
             if (process.platform === "win32") {
                 exec("taskkill", ["/PID", this._process.pid.toString(), '/T', '/F']);
